+++ client/src/components/RecipeDetailView.tsx
@@ -1,0 +1,12 @@
+import { useParams } from "react-router-dom";
+import { useRecipe } from "../app/recipes";
+import RecipeDetail from "./RecipeDetail";
+
+const RecipeDetailView = () => {
+    const { id } = useParams<{ id: string }>();
+    const recipe = useRecipe(id);
+
+    return recipe == null ? <p>Loading Recipe...</p> : <RecipeDetail recipe={recipe} />
+}
+
+export default RecipeDetailView;
